Bug in region drag&drop

+++ frontend/Selector.js
@@ -4,6 +4,7 @@ class Selector {
 
 		this.selected = null;
 		this.inDeleteMode = false;
+		this.typeToDelete = "";
 		this.editor = editor;
 		var lineLength = lineLength;
 
@@ -16,6 +17,7 @@ class Selector {
 
 		var offset = { x: 0, y: 0 };
 		var angleOffset = 0;  // angle offset for pose orientation (in radians)
+		this.selectedRegion = null;
 
 		function updateSelection(element) {
 			if (element.isSameNode(stage)) {
@@ -23,8 +25,8 @@ class Selector {
 				return;
 			}
 			var rect = element.getBoundingClientRect();
-			selection.style.left = rect.left + 'px';
-			selection.style.top = rect.top + 'px';
+			selection.style.left = (rect.left + window.pageXOffset) + 'px';
+			selection.style.top = (rect.top + window.pageYOffset) + 'px';
 			selection.style.width = rect.width + 'px';
 			selection.style.height = rect.height + 'px';
 			selection.style.display = 'block';
@@ -45,10 +47,22 @@ class Selector {
 					offset.x = parseFloat(target.getAttribute('cx')) - event.clientX;
 					offset.y = parseFloat(target.getAttribute('cy')) - event.clientY;
 				} else if (targetType === 'pose_line_annotation') {
-					offset.x = parseFloat(target.getAttribute('x1')) - event.clientX;
-					offset.y = parseFloat(target.getAttribute('y1')) - event.clientY;
+					var x1 = target.getAttribute('x1');
+					var y1 = target.getAttribute('y1');
+					offset.x = parseFloat(x1) - event.clientX;
+					offset.y = parseFloat(y1) - event.clientY;
+					angleOffset = Math.atan2(target.getAttribute('y2') - y1, target.getAttribute('x2') - x1);
+				} else if (targetType === 'region_annotation') {
+					// DEBUG
+					offset.x = parseFloat(target.parentElement.childNodes[1].getAttribute('cx')) - event.clientX; //event.clientX;
+					offset.y = parseFloat(target.parentElement.childNodes[1].getAttribute('cy')) - event.clientY; //event.clientY;
+					// offset.x = event.clientX;
+					// offset.y = event.clientY;
+					console.log(offset.x + " " + offset.y);
+				} else if (targetType === 'region_endpoint_annotation') {
+					offset.x = parseFloat(target.getAttribute('cx')) - event.clientX;
+					offset.y = parseFloat(target.getAttribute('cy')) - event.clientY;
 				}
-
 				self.selected = target;
 			}
 		});
@@ -75,6 +89,28 @@ class Selector {
 					}
 					self.selected.setAttribute('x2', x1 + lineLength * Math.cos(angleOffset));
 					self.selected.setAttribute('y2', y1 + lineLength * Math.sin(angleOffset));
+				} else if (targetType === 'region_annotation') {
+					// DEBUG
+					var translateStr = self.selected.parentElement.getAttribute('transform');
+					var translate = translateStr.substring(10, translateStr.length - 1);
+					var translateX = parseInt(translate.split(",")[0]);
+					var translateY = parseInt(translate.split(",")[1]);
+					var newOffsetX = event.clientX + offset.x - editor.getMidpointX();
+					var newOffsetY = event.clientY + offset.y - editor.getMidpointY();
+					this.console.log(newOffsetX + ", " + newOffsetY);
+					self.selected.parentElement.setAttribute('transform', 'translate(' + newOffsetX + ',' + newOffsetY + ')');
+				} else if (targetType === 'region_endpoint_annotation') {
+					// move the endpoint
+					var newX = event.clientX + offset.x;
+					var newY = event.clientY + offset.y;
+					self.selected.setAttribute('cx', newX);
+					self.selected.setAttribute('cy', newY);
+					// adjust the line
+					var currentRegion = self.selected.parentElement.childNodes[0];
+					var points = convertToList(currentRegion.getAttribute('points'));
+					var selectedEndpointId = parseInt(self.selected.getAttribute('id').split("-")[1]);
+					points[selectedEndpointId] = [newX, newY];
+					currentRegion.setAttribute('points', convertToString(points));
 				}
 				updateSelection(self.selected);
 			}
@@ -84,25 +120,45 @@ class Selector {
 			self.selected = null;
 		});
 
-		// DELETE
 		stage.addEventListener('click', function (event) {
 			var target = event.target;
 			if (target.isSameNode(stage) === false) {
-				if (self.inDeleteMode) {
-					var targetType = target.getAttribute('class');
+				var targetType = target.getAttribute('class');
+				if (self.inDeleteMode && targetType === self.typeToDelete) {
+					// DELETE
 					if (targetType === 'circle_annotation' || targetType === 'pose_line_annotation') {
 						self.editor.deleteElement(target);
 					}
+				} else if (targetType === 'region_annotation') {
+					self.enterRegionEditor(target);
 				}
 			}
 		});
 	}
 
-	enterDeleteMode() {
+	enterDeleteMode(typeToDelete) {
 		this.inDeleteMode = true;
+		this.typeToDelete = typeToDelete;
 	}
 
 	exitDeleteMode() {
 		this.inDeleteMode = false;
+		this.typeToDelete = "";
+	}
+
+	enterRegionEditor(target) {
+		window.document.getElementById("regionShapeBtns").style.display = "block";
+		$(':button:not(.regionShapeBtn)').prop('disabled', true);
+		this.selectedRegion = target;
+	}
+
+	exitRegionEditor() {
+		window.document.getElementById("regionShapeBtns").style.display = "none";
+		$(':button:not(.regionShapeBtn)').prop('disabled', false);
+		this.selectedRegion = null;
+	}
+
+	getSelectedRegion() {
+		return this.selectedRegion;
 	}
 }
